data_process/requirements.txt
google-api-python-client
google-cloud-pubsub
google-cloud-bigquery